Ensure that snapping respects vertial movement bounds.
This is because when moving PiP across displays, the PiP window could be released outside of the movement bounds and might get partially stuck on the upper or lower part of the display. This change ensures that the PiP window will always be within the vertical movement bounds after moving across displays.

Fixes: 422127560
Test: atest PipBoundsAlgorithmTest
Flag: com.android.window.flags.enable_dragging_pip_across_displays
Change-Id: Ia0f36665d2a04f290cacf1b033a75442f6e402e1