IME tracing: use perfetto protos
Adapt the input method tracing code to use the protos in
external/perfetto (SOT), instead of those in frameworks/base.

The protos in frameworks/base are a duplicate of the SOT that
still exist to ease the transition to perfetto, but eventually
they are going to be deleted.

Flag: EXEMPT mechanical refactoring
Test: presubmit
Bug: 276433199
Change-Id: Idc4fac363c6510520dd94e58a505ba2f4dbeca3f