Disallow stashing on cross-displays drag PiP window.
On cross-display drag, disallow stashing if display ID on down and on up
are different. Previously, this caused the PiP to be stashed on the edge
of the other display because the target display layout state bounds is
not updated until PiP is released.

Product decision to disallow stashing while dragging across displays.

Fixes: 423023871
Test: atest PipTouchHandlerTest
Flag: com.android.window.flags.enable_dragging_pip_across_displays
Change-Id: I8dc4abcd5dedbe3b840a7caaedb10cf69130856b