Correct bounds on moving PiP across displays.
Previously, we set the wct bounds to the bounds on release but before
snapping to the edge of the screen. This caused the PiP window to
misbehave, e.g. buttons not located correctly, stale PiP bounds on
locking/unlocking screen.

Fixes: 423685914
Fixes: 422370386
Fixes: 423999466
Test: manual and atest PipDisplayTransferHandlerTest
Flag: com.android.window.flags.enable_dragging_pip_across_displays
Change-Id: I25b6fe420063d2b1e1f6c36e4517d5f9529aff9f