Added unit tests for HsumBootUserInitializer constructor.
That class takes 2 boolean arguments, which in turn are defined by
a combination of multiple components (framework resources,
system properties, build type, and perhaps flags in the future).

This CL adds unit tests for the 2 methods that define those
arguments, and also move them to HsumBootUserInitializer itself
(previous they were defined on SystemServer, which is the class that
instantes HsumBootUserInitializer).

Bug: 402486365
Test: atest FrameworksMockingServicesTests --test-filter='.*HsumBootUserInitializer.*OnBootTest.*'
Flag: EXEMPT refactoring

Change-Id: I866d20b9356976e922168e41968a1a8dd56189f5