Fix multi-user bug with re-registered FRROs
On a device with multiple users (e.g., HSUM), when an existing FRRO is
re-registered (i.e., a new FRRO is created with same name/target/owner)
the resources for the target app are not updated with the new FRRO for
all users. In fact, only user 0 (system user on HSUM) gets the new
assets. For user 0, the new idmap is created but for later users, the
idmap is reused, hence OverlayManagerServiceImpl.updateState() returns
false and the target is not considered updated for the later users.

The fix is to also consider the baseCodePath (path to FRRO file) when
determining if a target is updated. The baseCodePath always changes when
there is a new FRRO.

Fixes: 406839614
Flag: EXEMPT bugfix
Test: atest OverlayManagerServiceImplTests
Test: atest FabricatedOverlayTests on HSUM device
Change-Id: I4417676ad165af8c149955efb59c52485771d278