Do not trigger IME task move on caption hold.
When a task is about to moved up due to IME, if a user is still holding
the caption down, do not move the task. This prevents a flicker that
happens when the transition is started but the user is dragging the
window around.

Test: Manual & atest DesktopImeHandlerTest
Fix: 415804441
Flag: com.android.window.flags.enable_desktop_ime_bugfix
Change-Id: Id3b1080cc69fd49419a3eed2224263214bde0c4d