Add debug logs to root cause drag test failure
DragAndDismiss test is still failing. From the video I can see that the
dismiss view is shown, but it does not react to the dragged bubble. And
the dragged bubble is not moving into it.
Adding some logs to track the instances of dismiss view itself and
magnetic target. They do get recreated. Want to confirm that the issue
is not from this.

Bug: 417749498
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.bubble.DragAndDismiss
Flag: EXEMPT, adding logging
Change-Id: I1210f631938731c4e6ee2cde79ce748cf8d62b30