Remove redundant sleep-token tracking
We ended up keeping multiple lists of sleep tokens
around which is wasteful and also makes it harder to
reason about how sleep management works.

At the end of the day, we just need a list of tokens per
display. This CL removes the extra lists and tracking.
Additionally, this revealed that the Acquirer was
re-entering the lock when it didn't need to. This is
important since cause/effect tracking is based on this
atomicity so now we can safely track causes of sleep/wake.

Bug: 423701190
Test: code-removal, so existing tests
Flag: EXEMPT: mechanical refactor
Change-Id: Iac6638ca217eb7f1c641d97b02bbd469e109b5fa