Prevent crash due to missing CUJ type.
Wear sysui is released from google3 and accesses jank monitor through reflection. This makes it harder to ensure we never install a apk that tries logging a CUJ that doenst exist on the platform yet due to it being an old build. If this ever happens, we should log the error instead of crashing.

Change-Id: I05cb432273b5cee5569b402f48cf2c9de3e07d30
Flag: NONE catching uncaught exception to reduce crashes in Wear
Bug: 429180834