System UI: Enable resizing edge for screen capture's region select box.
This change introduces drag to resize for the edges. We now have
ResizeZone, which is an interface that is one of Corner or Edge. For any
drag event, we will check whether a corner has been dragged first, and
then we check if an edge has been dragged.

Bug: 417533326
Test: Manual
Flag: com.android.systemui.desktop_screen_capture
Change-Id: Ia7b6d9cd4d6eeb949c3355c022d03650acfd07e5