Use immediate transactions to start/end early wakeup when surface control is not valid.
When closing the Shade instantly, this was failing because we were
trying to sync with a surface that was already gone. We can now do
it instantly as well without having to rely on that surface.

The only benchmark with the regression is ExpandNotificationByDraggingMicrobenchmark, that test:
 1. opens shade through AccessibilityService.GLOBAL_ACTION_NOTIFICATIONS a11y action
 2. sends a notification
 3. drags open the notification
 4. launches an app by tapping the notification.

The root cause is that the early wakeup wire-up in shade expansion a11y global action handler is before shade window is being made visible, so surfaceControl might not be valid at that point.
It could be that shade window wasn't hidden correctly before this flag when an app is launched from shade, with this flag that could have been fixed, which surfaced this regression.

Fixes: 423953389
Flag: com.android.systemui.instant_hide_shade
Test: unit test in child CL
Test: verified using a trace
Change-Id: I57aed26ff8fbeb568ba69fd52cbba50d1d244682