Add new call feature constant to support HD+ call.
[Description]
As of today, Dialer shows the HD+ icon in the Incall UI only,
when EVS codec is used.

The purpose of this feature is to extend the existing support
to also show the HD+ icon also in the call history.

[Changes]
Add a new feature flag constant to be set when the
call is considered a HD+ call.

Bug: 428115434
Flag: com.android.server.telecom.flags.hd_plus_call
Test: CTS verified
Change-Id: I8e9162f95da79baf95ade2a5acc11b776d4665e9