Ignore the Desk task when handling requestFullScreenMode
When validating the Activity#requestFullScreenMode API call to handle
the transition out of full-screen mode, the client controller was
checking that the top focused root task on the display was full screen
and could be restored.

In desktop mode however, this code was examining a special root task
container added by the organizer, and not the actual task for the
application, which was a child of it.

This change alters the code to obtain the task slightly. It now detects
the top-most task is the container task by checking the
mCreatedByOrganizer flag, and instead uses the topmost child task in the
the container. The code then proceeeds as before with that task.

Bug: 414334135
Flag: com.android.window.flags.enable_request_fullscreen_restore_freeform_bugfix
Test: atest CtsWindowManagerDeviceDisplay:FreeformWindowingModeTests
Change-Id: Ic0f97953b94734edc1e748ea1c026171f4738f6e