Fix GlonassAlmanac frequency channel range.
Frequency channel range should be -7..+6, not 0..31.

This is potentially a breaking change if the client mistakenly passes a negative number to the old API, which will lead to crashes. Therefore, the client must ensure check the flag fix_glonass_almanac_frequency_channel_range before using the API.

We are okay with this change because we are confident that no one has started to use this API yet.

Bug: 428050932
Bug: 358381377
Test: atest CtsLocationNoneTestCases
Flag: android.location.flags.fix_glonass_almanac_frequency_channel_range
Change-Id: I6cbe8ed011a70fd34e19a7d23920217b6fc850b8