Don't handle touch events if WindowDecoration is null
This CL adds null checks for the WindowDecoration for motion and click
listeners in DesktopModeWindowDecorViewModel. If the WindowDecor is
null, we shouldn't handle any more touch/hover/click events.

This is a short term fix for the issue where decoration.close triggers motion events, which when handled tries to get a reference to a WindowDecor that has been removed from the map (b/327664694). In the long term we should move event handling closer to the Decor.

Bug: 412749075
Test: atest DesktopModeWindowDecorViewModelTests
Flag: EXEMPT bugfix
Change-Id: I53b724234e238b86ae0220cdfeeb5651a6e8012a