Reset tint and override tint when the shelf expansion is collapsed
The issue: the shelf background can be transparent on the lockscreen in some circumstances.
Investigation: It's related to shelf expansion. After looking into the code base, we tint the shelf when expanding it into ExpandableNotificationRow(s). Since our notificationRowTransparency() logic in calculateBgColor is only active when tintColor and overrideTintColor are not set (to not override colorized notification colors), we are not able to reset the background color to opaque on the lockscreen in this case.
The solution: reset the tint and overrideTint colors after they are no longer used. However, please let me know if there is a better way to reset it.

Bug: 428615544
Flag: com.android.systemui.notification_row_transparency
Test: repeat the steps in the bug and make sure the background color of the shelf is correct.
Change-Id: I5f049dd8b5b00852f45e4cabcdcb513f65acf1a6