Update window context to use the application theme for the default theme
The issues in taskbar, magnification floating button and accessibility
floating menu come from the empty theme when the application calls
createWindowContext. When we perform getTheme to a context that has an
empty theme, we'll be led to the system default theme. In this change,
we modify the theme of the window context to the corresponding
application theme so the theme behaviors of the window context will
follow the application.

Bug: 406651629
Bug: 407669662
Bug: 407471267
Test: atest WindowContextTest
Flag: android.view.accessibility.force_invert_color
Change-Id: Ib0d9d2d3a4b26052a552f3e37cca6a143eb67a54