Set complete keyguard state change for new transition when going away
Going away will cause a transition to hide the keyguard. The reshow may
be blocked if the animation is cancelled before the pending lock is
handled. This will lead to a status mismatch between the system UI and
the WM.

Instead of calling to start a transition directly, go through the
setShowLocked() to avoid status mismatch.

Bug: 364930619
Test: KeyguardViewMediatorTest#testCancelKeyguardExitAnimationDueToSleep_withPendingLock_keyguardWillBeShowing
Flag: com.android.window.flags.ensure_keyguard_does_transition_starting_bug_fix
Change-Id: I759835c0765d9ddeffcc9bc787ae0c7728c233e1