Persist all of the repo state.
When there is a change in the repository, persist all of the desks
rather than just the desk that has had a change. This is needed because
back to back partial updates to the DataStore can create issues when
there are multiple removals.

Bug: 425916459
Test: atest DesktopRepositoryTest
Test: atest DesktopPersistentRepositoryTest
Flag: com.android.window.flags.repository_based_persistence

Change-Id: I57c0129d0ef65d357235beea6a9d77c6ec945d97