Go back to using LinearLayout in old MediaStyle
I17e6a805bb74bd4ad9ee9259ff6793ba5246570d changed all the layouts to use
a FrameLayout so that we could have the text on the second line in
collapsed notifications flow below the expander.

In the case of the old MediaStyle notifications, since it's possible for
action icons to be displayed on the right side when the notification is
collapsed, it ended up overlapping. Since for this particular style, the
actions are usually more important than the text, we'll go back to the
old layout to fix the issue.

Fix: 430601743
Test: manual, screenshot tests should be unchanged
Flag: android.app.notifications_redesign_templates
Change-Id: I189ab42631fad2b91c21aa6928302b4b31be41e5