Ensure talkback focus stays on maximize button after click
Fixing issue with clicking maximize button using Talkback. Often the talkback focus would go to the app chip instead of the maximize button. This seemed to be because of a race condition where the maximize button would be hidden, and talkback would foucs on the first focusable node of the app header instead. Now implementing a change utilizing the resize veil, to keep the focus and then manually giving it back to the maximize button after resizing.

Bug: 398975290
Flag: NONE bugfix
Test: manual (with talkback on)
Change-Id: I4a0b048c89be879f2c596ccfefef79522d3639eb