Prevent home activity being launched on a removed display
This CL fixes an issue where home activity could attempt to launch on a
display that has already been removed: this might happen if
DisplayManager has already removed the LogicalDisplay object, while
WindowManager has not yet received the `onDisplayRemoved()` event.

To prevent this, this CL adds a check of `Display#isValid()` in
`DisplayContent#isRemoved()` for non-default display.

Bug: 429409317
Test: DisplayContentTests
Flag: EXEMPT bugfix
Change-Id: I6290acc68228dd2427dfe786e3b5805cd7e6e654