Modify getShouldShowImeWithHardKeyboard return val
This modifies getShouldShowImeWithHardKeyboardForTesting to return a
nullable type, instead of the primitive boolean. The method is called
during a test teardown, before the IME process is actually stopped.
However, with config_preventImeStartupUnlessTextEditor set, the IME
process receives the onDestroy call as soon as the IME client is
removed. This would lead to a null pointer exception when calling this
getter method, as mSettingsObserver would already be null.

Flag: EXEMPT testfix
Bug: 406501026
Test: manually enable config_preventImeStartupUnlessTextEditor  and run
  atest InputMethodServiceTest#testImeSwitcherMenu_noLanguageSettingsWhenScreenLocked
Change-Id: I8a624268edf34c885627a094a06dbfaee5cd166b